Separate tests into archives_zuko
